+++ src/main/kotlin/io/github/jamalam/Bot.kt
@@ -79,5 +79,7 @@ suspend fun main() {
         }
     }
 
+    database.backup()
+
     bot.start()
 }